refactor(executors): 按厂商归并执行器实现
