+++ src/executors/anthropic/protocol.ts
