+++ src/executors/anthropic/sdk-trace.ts
