+++ src/executors/openai/api.ts
@@ -1,7 +1,7 @@
 import type { ExecResult, ExecutorInput } from '../../types/index.js';
 import { optionalTokenCount, splitInclusiveInputTokens } from '../../shared/token-usage.js';
 import { DEFAULT_TIMEOUT_MS } from '../core/defaults.js';
-import { readJsonResponse, responseBodyPreview } from './http.js';
+import { readJsonResponse, responseBodyPreview } from '../core/http.js';
 import { asErrorLike, errorMessage } from '../core/runtime.js';
 
 interface OpenAiResponse {
