+++ src/executors/openai/cli-trace.ts
