+++ src/executors/openai/protocol.ts
